% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.double2.R
\name{is.double2}
\alias{is.double2}
\alias{as.double2}
\title{\code{is.double()} for comma-decimal numbers}
\usage{
is.double2(x, dec = c(".", ","), na.rm = TRUE, ...)

as.double2(x, ...)
}
\arguments{
\item{dec}{characters to be treated as comma}

\item{na.rm}{ignore empty values}

\item{...}{currently not in use}
}
\description{
This works like \code{\link[=is.double]{is.double()}} and \code{\link[=as.double]{as.double()}}, but can also check (and transform) comma-decimal input such as \code{"0,1"}.
}
\details{
\code{is.double()} for comma-decimal numbers.
}
