% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\name{certedata_style_transformer}
\alias{certedata_style_transformer}
\title{Syntax Format According to Certe Coding Guidelines}
\usage{
certedata_style_transformer(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link[styler:tidyverse_style]{styler::tidyverse_style()}}}
}
\description{
Use this styler for formatting code by setting this as an option:

\code{options(styler.addins_style_transformer = "certedata_style_transformer()")}

Then use the keyboard shortcut of the \code{styler} package to apply the formatting.
}
\details{
The Certe styler keeps the first argument of a function on the same line, puts all following arguments on a new line, and does not add another new line after the last argument. This makes these lines:\preformatted{example_isolates \%>\% count(hospital_id, gender) \%>\% plot2(x.title = "Hospital", y.title = "Count", title = "Count isolates per hospital/gender")

example_isolates \%>\%
  count(hospital_id, gender) \%>\%
  plot2(x.title = "Hospital", y.title = "Count", title = "Count isolates per hospital/gender")
}

Change into:\preformatted{example_isolates \%>\%
  count(hospital_id,
        gender) \%>\%
  plot2(x.title = "Hospital",
        y.title = "Count",
        title = "Count isolates per hospital/gender")
}
}
