% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certepackage.R
\docType{package}
\name{certestyle-package}
\alias{certestyle}
\alias{certestyle-package}
\title{certestyle: A Certe R Package for Applying Certe Organisational Style}
\description{
A Certe R Package for applying the organisational colours and style of Certe, plus some additional formatting functions. This package is part of the 'certedata' universe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://certe-medical-epidemiology.github.io/certestyle}
  \item \url{https://github.com/certe-medical-epidemiology/certestyle}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.berends@certe.nl}

Authors:
\itemize{
  \item Erwin E. A. Hassing \email{e.hassing@certe.nl}
}

Other contributors:
\itemize{
  \item Certe Medical Diagnostics & Advice Foundation [copyright holder, funder]
}

}
\keyword{internal}
