% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourpicker.R
\name{colourpicker}
\alias{colourpicker}
\alias{as.character.colourpicker}
\alias{print.colourpicker}
\alias{add_white}
\title{Colours from \R, Certe, Viridis and More}
\usage{
colourpicker(x, length = 1, opacity = 0, ...)

\method{as.character}{colourpicker}(x, ...)

\method{print}{colourpicker}(x, ...)

add_white(x, white)
}
\arguments{
\item{x}{colour or colour palette name. Certe colours will be used from the \link{certe.colours} object. Input can be:
\itemize{
\item \code{"certe"}
\item \code{"certe0"} to \code{"certe6"} (higher numbers give lighter colours)
\item \code{"certeblauw"}, \code{"certegroen"}, \code{"certeroze"}, \code{"certegeel"}, \code{"certelila"}, or \code{"certebruin"} (or any of these followed by a 0 to 6)
\item \code{"certe_sir"} or \code{"certe_sir2"} for certeroze/certegeel/certegroen (will \strong{always} return length 5, with names "S", "SI", "I", "IR", "R")
\item One of the colourblind-safe \code{viridisLite} palettes:
\itemize{
\item \code{"viridis"}
\item \code{"magma"}
\item \code{"inferno"}
\item \code{"plasma"}
\item \code{"cividis"}
\item \code{"rocket"}
\item \code{"mako"}
\item \code{"turbo"}
}
\item One of the built-in palettes in \R (currently \R 4.5.1):
\itemize{
\item \code{"R3"}
\item \code{"R4"}
\item \code{"ggplot2"}
\item \code{"Okabe-Ito"}
\item \code{"Accent"}
\item \code{"Dark 2"}
\item \code{"Paired"}
\item \code{"Pastel 1"}
\item \code{"Pastel 2"}
\item \code{"Set 1"}
\item \code{"Set 2"}
\item \code{"Set 3"}
\item \code{"Tableau 10"}
\item \code{"Classic Tableau"}
\item \code{"Polychrome 36"}
\item \code{"Alphabet"}
\item \code{"topo"}
\item \code{"heatmap"}
\item \code{"rainbow"}
\item \code{"terrain"}
\item \code{"greyscale"}
\item \code{"grayscale"}
}
\item One of the 657 built-in \code{\link[=colours]{colours()}} in \R, such as \code{"brown2"}, \code{"lawngreen"}, \code{"magenta"}, \code{"snow4"}, \code{"wheat"}
}}

\item{length}{size of the vector to be returned}

\item{opacity}{amount of opacity (0 = solid, 1 = transparent)}

\item{...}{not used at the moment}

\item{white}{number between \verb{[0, 1]} to add white to \code{x}}
}
\value{
\link{character} vector in HTML format (i.e., \code{"#AABBCC"}) with new class \code{colourpicker}
}
\description{
Colours from \R, Certe, viridis and more. The output prints in the console with the actual colours.
}
\details{
Certe colours will be chosen as divergent as possible if the required output length is not too high. For example:
\itemize{
\item \code{x = "certe"} tries to only return the \code{"certe"} colours (\code{"certeblauw"}, \code{"certegroen"}, ...), the \code{"certe3"} colours (\code{"certeblauw3"}, \code{"certegroen3"}, ...) and the \code{"certe5"} colours (\code{"certeblauw5"}, \code{"certegroen5"}, ...)
\item \code{x = "certe2"} tries to only return the regular \code{"certe2"} colours (\code{"certeblauw2"}, \code{"certegroen2"}, ...), the \code{"certe4"} colours (\code{"certeblauw4"}, \code{"certegroen4"}, ...) and the \code{"certe6"} colours (\code{"certeblauw6"}, \code{"certegroen6"}, ...)
\item \code{x = "certe3"} tries to only return the \code{"certe3"} colours (\code{"certeblauw3"}, \code{"certegroen3"}, ...) and the \code{"certe5"} colours (\code{"certeblauw5"}, \code{"certegroen5"}, ...)
}

When using a single Certe colour with \code{length > 1}, a palette will be generated based on \*, \*2, \*3, \*4, and \*5. For example, \code{colourpicker("certeblauw", 5)} will return the colours \code{"certeblauw"}, \code{"certeblauw2"}, \code{"certeblauw3"}, \code{"certeblauw4"}, and \code{"certeblauw5"}.

A palette from \R will be expanded where needed, so even \code{colourpicker("R4", length = 20)} will work, despite "R4" only supporting a maximum of eight colours.
}
\examples{
colourpicker("certegroen")
colourpicker("certe", 5)
colourpicker(c("certeblauw", "red", "tan1", "#ffa", "FFAA00"))

par(mar = c(0.5, 2.5, 1.5, 0)) # set plot margins for below plots

# Certe colours
barplot(12:1,
        col = colourpicker("certe", 12),
        main = "'certe': uses 'certe' + 'certe3'")
barplot(12:1,
        col = colourpicker("certe2", 12),
        main = "'certe2': uses 'certe2' + 'certe4'")
barplot(12:1,
        col = colourpicker("certe3", 12),
        main = "'certe3': uses 'certe3' + 'certe5'")
barplot(5:1,
        col = colourpicker("certeblauw", 5),
        main = "'certeblauw'")
barplot(12:1,
        col = colourpicker("certeblauw", 12),
        main = "'certeblauw': auto-extended range")

# all colourblind-safe colour palettes from the famous viridisLite package
barplot(1:7,
        col = colourpicker("viridis", 7))
barplot(1:7,
        col = colourpicker("magma", 7))

barplot(8:1,
        col = colourpicker("R4", 8),
        main = "Some palettes have only 8 colours...")
barplot(20:1,
        col = colourpicker("R4", 20),
        main = "Not anymore!")


colours <- colourpicker("R4", 6)
colours
add_white(colours, 0.25)
add_white(colours, 0.5)
add_white(colours, 0.75)
}
