% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_mark.R
\name{dec_mark}
\alias{dec_mark}
\alias{big_mark}
\title{Use Decimal Comma?}
\usage{
dec_mark()

big_mark()
}
\description{
These functions determine which characters the decimal mark and big mark should be that are used in the Certe \R package functions. They base the determination on the \R \link[=Sys.getlocale]{locale user settings}.
}
\details{
For \code{\link[=dec_mark]{dec_mark()}}: this returns a comma (\code{","}) on Dutch systems, and a full stop (\code{"."}) otherwise. If the \link[=options]{option} \code{"dec_mark"} is set, that value will be used if it is either a comma or a full stop.

For \code{\link[=big_mark]{big_mark()}}: this returns a full stop if \code{\link[=dec_mark]{dec_mark()}} returns a comma, and a space otherwise. If the \link[=options]{option} \code{"big_mark"} is set, that value will be used if it is either a comma (\code{","}) or a full stop (\code{"."}) or a space (\code{" "}) or an empty character (\code{""}).
}
\examples{
# according the current user settings / OS language:
dec_mark()
big_mark()

options(dec_mark = ",")
dec_mark()
big_mark()

options(dec_mark = ".")
dec_mark()
big_mark()

options(big_mark = ",")
dec_mark()
big_mark()

# clean up
options(dec_mark = NULL, big_mark = NULL)
}
