% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_colours.R
\name{font_colours}
\alias{font_colours}
\alias{font_black}
\alias{font_blue}
\alias{font_green}
\alias{font_magenta}
\alias{font_red}
\alias{font_silver}
\alias{font_white}
\alias{font_yellow}
\alias{font_subtle}
\alias{font_grey}
\alias{font_bold}
\alias{font_italic}
\alias{font_underline}
\alias{font_stripstyle}
\title{Console Font Colours}
\usage{
font_black(..., collapse = " ")

font_blue(..., collapse = " ")

font_green(..., collapse = " ")

font_magenta(..., collapse = " ")

font_red(..., collapse = " ")

font_silver(..., collapse = " ")

font_white(..., collapse = " ")

font_yellow(..., collapse = " ")

font_subtle(..., collapse = " ")

font_grey(..., collapse = " ")

font_bold(..., collapse = " ")

font_italic(..., collapse = " ")

font_underline(..., collapse = " ")

font_stripstyle(...)
}
\arguments{
\item{...}{character (vector)}

\item{collapse}{character to separate the text elements. Use \code{collapse = NULL} to \emph{not} collapse the input.}
}
\description{
Add colours and font formatting to console text.
}
\examples{
cat(font_black("TEXT TEST"), "\n")
cat(font_red("TEXT TEST"), "\n")
cat(font_green("TEXT TEST"), "\n")
cat(font_blue("TEXT TEST"), "\n")
cat(font_bold("TEXT TEST"), "\n")
cat(font_italic("TEXT TEST"), "\n")
cat(font_underline("TEXT TEST"), "\n")
}
