% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plain_html_table.R
\name{plain_html_table}
\alias{plain_html_table}
\title{Format Data Set as HTML}
\usage{
plain_html_table(x, max_col = Inf)
}
\arguments{
\item{x}{a \link{data.frame}}

\item{max_col}{maximum number of columns to return}
}
\description{
Formats a \link{data.frame} into HTML code, so it can be used in e.g. \code{\link[certemail:mail]{mail()}} or \code{\link[certeprojects:teams]{teams()}}.
}
\examples{
plain_html_table(mtcars[1:2, 1:2])
}
