% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.data.frame.R
\name{print.data.frame}
\alias{print.data.frame}
\title{Printing Data Frames}
\usage{
\method{print}{data.frame}(x, ...)
}
\arguments{
\item{x}{object of class \code{data.frame}.}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
Print a data frame.
}
\details{
This calls \code{\link[base]{format}} which formats the data frame
  column-by-column, then converts to a character matrix and dispatches
  to the \code{print} method for matrices.

  When \code{quote = TRUE} only the entries are quoted not the row names
  nor the column names.
}
\examples{
(dd <- data.frame(x = 1:8, f = gl(2,4), ch = I(letters[1:8])))
     # print() with defaults
print(dd, quote = TRUE, row.names = FALSE)
     # suppresses row.names and quotes all entries
}
\seealso{
\code{\link[base]{data.frame}}.
}
