% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{rmarkdown}
\alias{rmarkdown}
\alias{rmarkdown_author}
\alias{rmarkdown_date}
\alias{rmarkdown_table}
\alias{rmarkdown_template_file}
\alias{rmarkdown_template_folder}
\title{Markdown Template Properties}
\usage{
rmarkdown_author(user_id = NULL)

rmarkdown_date(date = Sys.Date())

rmarkdown_table(...)

rmarkdown_template_file()

rmarkdown_template_folder()
}
\arguments{
\item{user_id}{user ID at Certe, defaults to currently logged in user}

\item{date}{Date to print in "d mmmm yyyy"}

\item{...}{data set to print using \code{\link[certetoolbox:tbl_flextable]{certetoolbox::tbl_flextable()}}}
}
\description{
These functions can be used in R Markdown documents.
}
