% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toproper.R
\name{toproper}
\alias{toproper}
\title{Proper Case for Text}
\usage{
toproper(text, every_word = FALSE, ...)
}
\arguments{
\item{text}{text to transform}

\item{every_word}{set captial to every word}
}
\description{
Proper Case for Text
}
