% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.double2.R
\name{is.double2}
\alias{is.double2}
\alias{as.double2}
\title{\code{is.double()} for Comma-Decimal Input}
\usage{
is.double2(x, dec = c(".", ","))

as.double2(x, dec = c(".", ","))
}
\arguments{
\item{x}{vector of values}

\item{dec}{characters to be treated as comma}
}
\description{
This works like \code{\link[=is.double]{is.double()}} and \code{\link[=as.double]{as.double()}}, but is vectorised and can also check (and transform) comma-decimal input such as \code{"0,1"}.
}
\examples{
is.double(0.1)
is.double("0.1")
is.double("0,1")

is.double2(0.1)
is.double2("0.1")
is.double2("0,1")

is.double(c(0.1, "0.1", "0,1"))
is.double2(c(0.1, "0.1", "0,1"))

as.double(c(0.1, "0.1", "0,1"))
as.double2(c(0.1, "0.1", "0,1"))
}
