% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certe_colours.R
\docType{data}
\name{certe.colours}
\alias{certe.colours}
\title{Certe Colour Vector}
\format{
An object of class \code{character} of length 42.
}
\usage{
certe.colours
}
\description{
This is a character vector with all Certe colours. It is used by \code{\link[=colourpicker]{colourpicker()}}.
}
\details{
The colours are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#>      certeblauw      certegroen       certeroze       certegeel       certelila 
#>       "#4A647D"       "#93984C"       "#B4527F"       "#E4D559"       "#825988" 
#>  certezachtlila     certeblauw2     certegroen2      certeroze2      certegeel2 
#>       "#CEB9D6"       "#69849C"       "#B8BB8B"       "#C794AC"       "#E4DD9C" 
#>      certelila2 certezachtlila2     certeblauw3     certegroen3      certeroze3 
#>       "#AE92B1"       "#DED3E2"       "#7A93A9"       "#C9CCA5"       "#D5ACBF" 
#>      certegeel3      certelila3 certezachtlila3     certeblauw4     certegroen4 
#>       "#ECE6B1"       "#C1ABC4"       "#E6DDE9"       "#C5D0DB"       "#DFE1C8" 
#>      certeroze4      certegeel4      certelila4 certezachtlila4     certeblauw5 
#>       "#E7CCD8"       "#F5F1CF"       "#DACCDC"       "#F0EAF2"       "#E2E7EC" 
#>     certegroen5      certeroze5      certegeel5      certelila5 certezachtlila5 
#>       "#EEEFE4"       "#F2E6EB"       "#F9F7E8"       "#ECE6ED"       "#F7F4F8" 
#>     certeblauw6     certegroen6      certeroze6      certegeel6      certelila6 
#>       "#F6F7F8"       "#F9F9F6"       "#FAF7F8"       "#FCFBF8"       "#F8F6F8" 
#> certezachtlila6     certeblauw0     certegroen0      certeroze0      certegeel0 
#>       "#FCFBFC"       "#3A4D5D"       "#5A5D33"       "#7F3C5B"       "#D4C230" 
#>      certelila0 certezachtlila0 
#>       "#503852"       "#BEA5C7"
}\if{html}{\out{</div>}}
}
\keyword{datasets}
