% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourpicker.R
\docType{data}
\name{certe.colours}
\alias{certe.colours}
\title{Certe Colour Vector}
\format{
An object of class \code{character} of length 42.
}
\usage{
certe.colours
}
\description{
This is a character vector with all Certe colours.
}
\examples{
certe.colours
}
\keyword{datasets}
