% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourpicker.R
\name{colourpicker}
\alias{colourpicker}
\alias{as.character.colourpicker}
\alias{print.colourpicker}
\title{Colours from \R, Certe, Viridis and More}
\usage{
colourpicker(x, length = 1, opacity = 0, ...)

\method{as.character}{colourpicker}(x, ...)

\method{print}{colourpicker}(x, ...)
}
\arguments{
\item{x}{colour or colour palette name. Certe colours will be used from the \link{certe.colours} object. Input can be any colour from base \R \code{\link[=colours]{colours()}}, or:
\itemize{
\item \code{"certe"}
\item \code{"certe0"} until \code{"certe6"} (higher numbers give lighter colours)
\item \code{"certeblauw"}, \code{"certegroen"}, \code{"certeroze"}, \code{"certegeel"}, \code{"certelila"}, or \code{"certezachtlila"} (or any of these followed by a 0 until 6)
\item \code{"certe_rsi"} or \code{"certe_rsi2"} for certeroze/certegeel/certegroen (will \strong{always} return length 3)
\item \code{"R3"} for colours from \R version 3
\item \code{"R4"} for colours from \R version 4
\item \code{"viridis"} for colourblind-safe colours
\item \code{"topo"}
\item \code{"heatmap"}
\item \code{"rainbow"}
\item \code{"terrain"}
\item \code{"greyscale"} or \code{"grayscale"}
}}

\item{length}{size of the vector to be returned}

\item{opacity}{amount of opacity (0 = solid, 1 = transparent)}

\item{...}{not used at the moment}
}
\value{
\link{character} vector in HTML format (i.e., \code{"#AABBCC"}) with new class \code{colourpicker}
}
\description{
Colours from \R, Certe, viridis and more. The output prints in the console with the actual colours.
}
\examples{
colourpicker("certegroen")
colourpicker("certe", 5)
colourpicker(c("certeblauw", "red", "tan1", "ffa", "#FFAA00"))
}
