% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{current_markdown_colour}
\alias{current_markdown_colour}
\title{Get Current Markdown Colour}
\usage{
current_markdown_colour()
}
\description{
This function determines the Certe theme colour currently used in a markdown document (Quarto or R Markdown), based on the YAML header.
}
\details{
It returns a Certe colour if one is set in the YAML header, and checks in this order:
\enumerate{
\item \code{colour-main}
\item \code{colour-titlepage-titlebanner}
\item \code{colour-titlepage-full}
\item \code{colour-verticalbars}
\item \code{colour-heading1}
}

If none is set, it returns \code{"certeblauw"}.

This function is the default to set the theme for \code{\link[certetoolbox:tbl_flextable]{tbl_flextable()}}.
}
