% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format2.R
\name{format2}
\alias{format2}
\alias{format2.default}
\alias{format2.numeric}
\alias{format2.percentage}
\alias{format2.percent}
\alias{format2.Date}
\alias{format2.POSIXt}
\alias{format2.hms}
\alias{format2.difftime}
\alias{format2_scientific}
\title{Formatting with readable \code{format} settings and Dutch defaults}
\usage{
format2(x, ...)

\method{format2}{default}(
  x,
  format = "d mmmm yyyy",
  percent = FALSE,
  round = ifelse(percent, 1, 2),
  force.decimals = ifelse(percent, TRUE, FALSE),
  decimal.mark = ",",
  big.mark = ".",
  locale = "nl",
  ...
)

\method{format2}{numeric}(
  x,
  round = ifelse(percent, 1, 2),
  force.decimals = ifelse(percent, TRUE, FALSE),
  decimal.mark = ",",
  big.mark = ".",
  min.length = 0,
  percent = FALSE,
  ...
)

\method{format2}{percentage}(x, round = NULL, decimal.mark = ",", big.mark = ".", ...)

\method{format2}{percent}(...)

\method{format2}{Date}(x, format = "d mmmm yyyy", locale = "nl", ...)

\method{format2}{POSIXt}(x, format = "HH:MM:SS", locale = "nl", ...)

\method{format2}{hms}(
  x,
  format = "HH:MM:SS",
  round = 2,
  force.decimals = FALSE,
  decimal.mark = ",",
  big.mark = ".",
  ...
)

\method{format2}{difftime}(
  x,
  round = 2,
  force.decimals = FALSE,
  decimal.mark = ",",
  big.mark = ".",
  ...
)

format2_scientific(x, decimal.mark = ",", ...)
}
\arguments{
\item{x}{vector of values}

\item{...}{arguments given to methods such as \code{\link[=format]{format()}}}

\item{format}{format to use, can be set with human-readable text such as \code{"d mmmm yyyy"} or POSIX such as \code{"\%e \%B \%Y"}}

\item{percent}{logical to transform numeric to percentage (character)}

\item{round}{number of decimals to round to}

\item{force.decimals}{force printing decimals, even with trailing zeroes}

\item{decimal.mark, big.mark}{decimal and thousands limiters}

\item{locale}{language to set for dates}

\item{min.length}{minimal length of output, overwrites \code{force.decimals}}
}
\value{
\code{\link[=format2]{format2()}} always returns a \link{character}.
}
\description{
Formatting with readable \code{format} settings and Dutch defaults
}
\details{
The \code{\link[=format2_scientific]{format2_scientific()}} function returns an \link{expression} and can be used in \code{ggplot2} plots.
}
\examples{
format2("2021-01-01")

format2(Sys.time(), "d mmmm yyyy HH:MM")

# content-aware of decimal use
format2(1024)
format2(c(1024, 0.123))

format2(2.1)
format2(2.1, force.decimals = TRUE) # since default is 2 decimals

p <- cleaner::as.percentage(0.123)
format2(p)

# use format2_scientific for scientific labels in plots:
# if (require("certeplot2")) {
#    plot2(mtcars,
#          y = hp * 1000,
#          y.labels = format2_scientific)
# }
if (require("ggplot2")) {
  ggplot(mtcars) +
    geom_point(aes(x = mpg, y = hp * 1000)) +
    scale_y_continuous(label = format2_scientific)
}
}
