% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{rmarkdown}
\alias{rmarkdown}
\alias{rmarkdown_author}
\alias{rmarkdown_department}
\alias{rmarkdown_date}
\alias{rmarkdown_table}
\alias{rmarkdown_template}
\alias{rmarkdown_logo}
\title{Markdown Template Properties}
\usage{
rmarkdown_author(user_id = NULL)

rmarkdown_department()

rmarkdown_date(date = Sys.Date())

rmarkdown_table(...)

rmarkdown_template(type = "latex")

rmarkdown_logo(logo_type = "front")
}
\arguments{
\item{user_id}{user ID at Certe, defaults to currently logged in user}

\item{date}{Date to print in "d mmmm yyyy"}

\item{...}{data set (and options) to pass on to \link[certetoolbox:tbl_flextable]{tbl_flextable()}}

\item{type}{defaults to "latex", must be "latex" or "word" (case-insensitive)}

\item{logo_type}{type of logo, must be "front" or "footer" (case-insensitive). For the LaTeX template, the front logo must be max 16x7 cm, and the footer logo must be max 16x0.7 cm.}
}
\description{
These functions can be used in R Markdown documents.
}
\details{
For LaTeX, 4 logo files can be used to add to the resulting PDF file: \code{logocentre} (38x58 mm), \code{logoleft} (38x58 mm), \code{logoright} (38x58 mm) and \code{logofooter} (7 mm high). They must be set as YAML parameters in an R Markdown file header.
}
