% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_themes.R
\name{rstudio_install_certe_themes}
\alias{rstudio_install_certe_themes}
\alias{rstudio_set_certe_light}
\alias{rstudio_set_certe_dark}
\title{Certe Themes for RStudio}
\usage{
rstudio_install_certe_themes()

rstudio_set_certe_light()

rstudio_set_certe_dark()
}
\description{
Install and apply RStudio syntax highlighting in Certe theme colours.
}
\details{
This package comes with two styles that can be installed with \code{\link[=install_rstudio_certe_themes]{install_rstudio_certe_themes()}}: "Certe Light" and "Certe Dark". Needless to say, this function requires RStudio.

Quickly apply the light theme with \code{\link[=rstudio_certe_light]{rstudio_certe_light()}} and the dark theme with \code{\link[=rstudio_certe_dark]{rstudio_certe_dark()}} (they will be installed if needed).
}
